Refactored network manager to do all the handling of requests Made onSuccess of network respose to just return a JSONObject or exception.
