+++ UWCalendar/src/main/AndroidManifest.xml
@@ -10,7 +10,7 @@
         android:label="@string/app_name"
         android:theme="@style/AppTheme" >
         <activity
-            android:name=".activities.SplashScreenActivity"
+            android:name=".activities.SubjectsActivity"
             android:label="@string/app_name"
             android:configChanges="orientation|screenSize" >
             <intent-filter>
