+++ UWCalendar/src/main/java/com/alirezatr/uwcalendar/listeners/BaseResponseListener.java
@@ -1,0 +1,8 @@
+package com.alirezatr.uwcalendar.listeners;
+
+import org.json.JSONObject;
+
+public interface BaseResponseListener {
+    void onSuccess(JSONObject response);
+    void onError(Exception e);
+}
